refactor: decouple sidebar to accept sections and actions from caller
